Reorganize QDaqTypes.h: Create QDaqVector.h as a cpp-less class in the style of template.
